+++ lib/loader/config_loader.js
@@ -10,12 +10,12 @@ const extend = require('extend2');
 
 const run = (appInfo) => {
   // 載入框架預設的 config
-  const mahudasConfigPath = path.join(path.dirname(__dirname), './config/config.default.js');
+  const mahudasConfigPath = path.join(path.dirname(__dirname), '../config/config.default.js');
   const mahudasDefaultConfig = require(mahudasConfigPath);
 
   const appRoot = appInfo.root;
   // 載入config.default
-  let configFilePath = path.join(appRoot, './config/config.default.js');
+  let configFilePath = path.join(appRoot, '../config/config.default.js');
   let defaultConfig = {};
   if (fs.existsSync(configFilePath)) defaultConfig = require(configFilePath);
