+++ lib/loader/core_middleware_config_loader.js
