+++ lib/loader/service_loader.js
@@ -43,6 +43,9 @@ const recursiveSearch = async (startPath, subPath = '') => {
     const scanPath = path.join(startPath, subPath);
     const directX = fs.readdirSync(scanPath, { withFileTypes: true });
     await async.eachSeries(directX, async (dirent) => {
+      // 如果命名是"-"開頭，就忽略
+      if (dirent.name.indexOf('-') === 0) return;
+
       const currentPath = path.join(scanPath, dirent.name);
       if (dirent.isDirectory()) {
         const subModules = await recursiveSearch(startPath, path.join(subPath, dirent.name));
